+++ CodeConverter/CodeWithOptions.cs
@@ -20,11 +20,7 @@ public class CodeWithOptions
 
     public string Text { get; private set; }
     public string FromLanguage { get; private set; }
-    [Obsolete("This option has no effect. Please use a code formatter to format the output code.")]
-    public int FromLanguageVersion { get; private set; }
     public string ToLanguage { get; private set; }
-    [Obsolete("This option has no effect. Please use a code formatter to format the output code.")]
-    public int ToLanguageVersion { get; private set; }
 
     public IReadOnlyCollection<PortableExecutableReference> References { get; set; } = new List<PortableExecutableReference>();
 
@@ -42,20 +38,6 @@ public class CodeWithOptions
     }
 
     public CodeWithOptions SetToLanguage(string name = LanguageNames.VisualBasic)
-    {
-        ToLanguage = name;
-        return this;
-    }
-
-    [Obsolete("The version option has no effect, the latest version will be used.")]
-    public CodeWithOptions SetFromLanguage(string name, int version)
-    {
-        FromLanguage = name;
-        return this;
-    }
-
-    [Obsolete("The version option has no effect, the latest version will be used. Please use a code formatter to format the output code.")]
-    public CodeWithOptions SetToLanguage(string name, int version)
     {
         ToLanguage = name;
         return this;